module/requirements.txt
transformers
umap-learn
compress-fasttext
nltk
pandas
sklearn
tqdm
matplotlib
torch